refactor: convert to new system and add multi-platform support

+++ uninstall.sh
@@ -1,9 +1,15 @@
 #!/usr/bin/env bash
 
-choco.exe uninstall vcxsrv -y
+SYSTEM=$3
 
-WIN_STARTUP_PATH=$(powershell.exe -Command "[Environment]::GetFolderPath(\"Startup\")")
-STARTUP_PATH=$(wslpath "$WIN_STARTUP_PATH")
-STARTUP_PATH=${STARTUP_PATH%$'\r'}
+if [ "$SYSTEM" == "wsl" ]; then
+    choco.exe uninstall vcxsrv -y
 
-rm "$STARTUP_PATH/config.xlaunch"
+    WIN_STARTUP_PATH=$(powershell.exe -Command "[Environment]::GetFolderPath(\"Startup\")")
+    STARTUP_PATH=$(wslpath "$WIN_STARTUP_PATH")
+    STARTUP_PATH=${STARTUP_PATH%$'\r'}
+
+    rm "$STARTUP_PATH/config.xlaunch"
+else
+    echo -e "\e[33mVcxSrv is only installable on Windows platforms, nothing to uninstall."
+fi
